rag-quran-server/requirements.txt
flask==3.0.3
flask-cors==5.0.1
python-dotenv==1.0.1
faiss-cpu==1.9.0.post1
pandas==2.2.3
google-generativeai==0.7.2
sentence-transformers==3.4.1
